Allow logging exceptions to propagate normally
By default, exceptions in python's logging module are not propagated.
This means that, for example, logging format bugs in the main code will
not cause the program to exit, thereby making it harder to notice them.

This patch adds a custom logging StreamHandler that overrides the
default exception handling to print the exception to stderr, and then
re-raise the exception.

https://docs.python.org/3/howto/logging.html#exceptions-raised-during-logging