Don't auto-restart processes which crash
If supervisor instantly restarts a crashed process, it can be hard to
notice. As we're in heavy development, it's far better to have a clear
signal that a process has crashed (and is clearly shown in the output of
supervisor.sh status)

Also update the log filenames slightly to be consistent with the
supervisor process names. (I.e. ad2-21:bs2-21-1 will now log to
bs2-21-1.log, and not bs-2-21-1.log).