Remove IP addresses from commandlines
Building on rev112's work, the infrastructure servers now get an ID on
the commandline, and then look up their specific config in the topology
file. It gets rid of redundant information (IP on commandline and in
topology file), allows each server to know their own identity, and means
i can remove the depdancy on the special supervisor env var.