Change ZKSharedCache to not rely on timestamps.
Due to ZK consistency issues, it's possible for multiple entries with
the same timestamp to become visible at different times. This means that
using the latest timestamp of a processed entry to discard all older
entries on the next read can cause updated entries to be missed.

Switching instead to unique entry names allows us to determine just on
the children list which entries are new, no need to fetch the metadata
of every entry each run.

Instead of caching the zk timestamp of an entry, we now cache the
timestamp of the first time we read (or wrote) the entry. This means
that expiry is very simple, and doesn't even need to query zk to get a
list of the current children. Anything old, we already know about;
anything new doesn't need to be expired.

By storing the timestamp on write, it also allows us to avoid reading
PCBs that we have written, so this change also adds a queue to the
beacon server so it can process incoming PCBs directly, instead of
waiting for them to be read back from ZK.

(Also fix an issue with assert_these_calls where any_order wasn't being
passed through.)