Use bin-packing(-like) algorithm for ip allocation.
The allocation algorithm hands out place-holder (AddressProxy) objects
on the first pass through the topology, then when the total number of
elements per subnet is known, goes back and assigns subnets and
addresses to each element.

A custom JSONEncoder class then converts the proxy objects to strings at
the point where the topology is written to a JSON file on disk. This is
mostly transparent to the rest of the code.

Also:
- Switch to storing interfaces rather than addresses in topology files
  (i.e. "127.0.0.217/29" rather than "127.0.0.217"). This will make it
  easier for an updated web management interface to know where to
  allocate new IPs.
- Always write out the network configuration (renamed from
  "mininet.conf" to "network.conf").
- Remove "AddrType" from topology files. The code can now figure out the
  type itself (and this simplifies the ip allocation code quite a bit).
- Refactored lib.topology.parse_dict to be testable/readable.
- Delete /run/shm/scion-zk when generating topology, as the ZK data dirs
  are already being wiped, and having the stale datalogs makes ZK crash
  on startup.
- The topology generator is now completely ip-version agnostic.