Generate per-element topology/configs.
The generated directory structure looks like this:
gen/
    ISD1/
        AD10/
            bs1-10-1/       # Configuration for bs1-10-1 element
                ad.conf
                path_policy.conf
                supervisord.conf
                topology.conf
                certs/      # Certs for AD1-10
                    ISD1-V0.crt
                    ISD1-AD10-V0.crt
                keys/
                    ad-sig.key  # Signing key for AD1-10
            common/    # Common ad.conf and topology.conf

Each element has its own fully independant config directory, with copies
of the AD's configs/certs/keys, and a supervisord.conf that lists only
the element. This allows easy deployment of an element on a machine.

There's a common/ subdir at the AD level, for use by sciond and anything
else that needs access to AD-wide config.

Also:
- Make all servers only require a server id, and an optional config dir
  argument, on the command line.
- Unify main() for all servers.
- Improved cert chain request/reply logging
- Simplify Topology.get_own_config, and improve error handling.