Improve POX detection robustness
We used to sleep for 1s and assume that POX would have sufficient time
to load, but on a fresh boot it might take longer than 1s. This new way
of doing things sleeps at 1s intervals up to a max of 5s, checking at
each interval whether POX has actually started. If the max time is
reached, the script aborts and displays the POX debug output.

Thanks @kormat for the help