improve pox detection robustness
we used to sleep for 1s and assume that pox would load, but it turns out on a fesh boot it might take longer than that. this new way of doing it sleeps as long as pox hasn't loaded, which ensures pox is always running prior to doing mininet things.