Switch from json to yaml for human-generated configs.
JSON is very human-hostile for creation/editing. It is unforgiving of
missing (or extra) commas, requires braces, and is generally difficult
to work with. YAML is much more compact, and much easier for humans to
edit. This is a major benefit when trying to create topologies, for
example. YAML is also a (mostly) drop-in replacement for JSON.

Also:
- Move config file names into lib.defines
- Remove unused topologies (shoelace4 and MultiPathTest), which are
  unmaintained, and not compatible with the recent generator changes.