Allow clients to query for paths to any core AS
By specifying AS 0, a client can ask for a path to any core AS. If the
specified ISD is the local ISD, this will return up segments only. If
it's a remote ISD, then it will get both up and core segments.

This required expanding lib.requests slightly, so that programs that use
it can specify what answers satisfy which requests. This allows sciond
to tell lib.requests that ISD,0 requests can be fulfilled by any ISD,X
response.