Small improvements to dependency installer script
By using the set -e directive, the script would exit mysteriously if any of the
invoked commands returns non-zero status. This happens, for example when apt
tries to restart zookeperd or ntp. By not having set -e the script will try to
go through all the steps in the script. This also now shows the help when
deps.sh is run without arguments.

Also modify the wording slightly to better reflect what each option does.
Previous helped appeared to give the user the option of installing packages with
or without root. Hopefully it is now clear that all packages need to be
installed, but only some require root privs to install. It was not clear what
misc meant, when it currently only installs supervisor packages. Renamed the
functions accordingly.