Allow extension handlers to direct packet processing
Extension handlers can return a list RouterFlags (with arguments,
depending on the flag), that the router can then process to see what
needs to happen to the packet.

An example use is most SIBRA packets won't use the normal path
processing in the router, and will instead be forwarded directly
instead.

Also:
- Fix minor log formatting in path server
- Identify whether it's a pre- or post- handler that's missing
- Print out a hexdump of a packet if payload parsing failed, otherwise
  this can trigger further exceptions.