Make integration tests much less noisy.
With the huge amounts of text the integration tests are currently
outputting, browsers looking at circleci output become unresponsive.
Almost all of it is noise in any case, so this PR changes all of the
non-essential logging to DEBUG, and sets the default console level to
INFO. All of the information is still available in the DEBUG logs, it's
just not printed to the console.