Start switching payload serilization to capnproto.
Most of the changes are related to the fact that a capnp object should
not be packed multiple times (as this can cause memory leakage, or
wasted data bytes). Currently we pack() everything multiple times to
calculate lengths. This change stops that (and has an assertion to make
sure it won't happen in future). One caveat is that packets that have
been assembled (as opposed to created) won't display meaningful lengths.
E.g:
- the total_len field in the common header isn't populated during packet
  assembly, and is instead filled in during packet packing.
- UDP and SCMP checksums aren't calculated before packet packing.

This does not apply to parsed packets.

Details:
- Install pycapnp from pip to provide capnproto python support.
- Install the current capnp library/tools release.
- Created SCIONPayloadBaseProto to be the new base class for any capnp
  payload classes.
- IFIDPayload is now a capnp message.
- Change sphinx building around so that it builds in the root directory,
  otherwise the .capnp schema loading will fail.