Use github.com/rancher/trash to manage certain Go dependencies.
Dealing with forked projects in Go is somewhat painful. You either need
to change all relevant import lines (including those in the dependency
itself), or check out the fork into the location Go expects the original
project to be. By using trash, we can simplify this a lot. It reads the
vendor.conf in go/, and checks out the desired revision of the
dependency from a specified repo, which can be a fork.