Fix mutex locking in SSP and related code
To debug correct use of mutexes in multithreaded applications, it's always hand
yo use lock annotations:

http://clang.llvm.org/docs/ThreadSafetyAnalysis.html

In order to do this in our C++ code, this commit does the following:

1. Make a wrapper around pthread mutexes that is annotated with the correct
   macros (`ACQUIRE()`, `RELEASE()`, `CAPABILITY("mutex")` and so on) (in
   `MutexScion.h`)
2. Import the preprocessor macros that are used by the Clang thread safety
   analysis tooling (`Mutex.h`, taken from the Apache-2 licensed lmctfy).
3. Change all uses of mutex-reladed pthread functions to use the class from
   (1). There are some raw uses of pthread functions left in
   `SCIONWrapper.h`/`.cpp` since that code is used with `extern "C"`. Said code
   is simple enough to be checked manually.
4. Annotate all relevant C++ functions/methods with `EXCLUDES()` directives.
5. Compile the code with `-Wthread-safety`.

This leads to the discovery of various `threadCleanup()` methods that release
mutexes "just in case". While this sometimes works well enough, double releases
of mutexes trigger undefined behavior and thus can lead to SEGV and a host of
other problems. On top of that dealing with locks in this manner can easily
paper over concurrency problems.

Note that all of this annotation is for compile-time checks _only_. Thus, it
should have no performance or memory impact after compilation.

This commit also adds a wrapper around `fprintf()` (`debugprint` in
`Utils.h/cpp`) that allows for easily squelched debug messages to be printed to
stderr.

Finally, this commit adds appropriate file header comments containing the ETH
copyright notice and a reference to the Apache-2 license.