Deprecate libssocket. (#927)
libssocket (./endhost/ssp) has major threading/locking issues. These
prevent it from working at all on the new HP DL20 servers we're starting
to buy. Our best guess is that this is because the servers have
non-hyperthreaded CPUs, and this is somehow triggering otherwise latent
bugs in the thread handling.

Fixing this would involve a major amount of engineering time (and
probably a rewrite). So, this change stops building/testing libssocket
by default. It can still be used as a proof of concept, but is
definitely unsuitable for any kind of production traffic for the time
being.