Rename go/border/packet.Packet to go/border/rpkt.RPkt (#929)
The Go router's Packet type is very tightly bound to an underlying byte slice, for performance reasons. This is great for high-speed processing, but becomes a large obstacle when more fundamental changes to the packet are desired (e.g. reversing the packet).

In preparation for making a more general scion packet type (SPkt), this change renames the router's packet.Packet to rpkt.RPkt (for "Router packet").