Fix packet recycling with libhsr in the go router. (#943)
readHSRInput was changed to keep a static set of packet buffers for
re-use recently, but the packets were still recycled, meaning they were
available for other parts of the router to reuse. This leads to an
interesting and wide variety of strange errors. This change resets the
packet buffers instead, which clears the contents.

Also:
- Fix a ~cosmetic with how hsr.AddrMs is updated that made debugging
  much harder than it needed to be. The old version takes a /copy/ of
  each element of AddrMs, and modifies the copy, leaving the original
  untouched. The new version fixes that.
- Remove the VSS ulimit, as it causes the hsr library to die fatally on
  startup without much useful information.