Add libcapnp-dev to packages list
This greatly reduces the amount of work that the pycapnp installation
has to do:

With 500M of ram, it takes 3.5mins for pycapnp to install without the
headers, and 35s with the headers. With less than 500M of ram, pycapnp
install will fail without the headers, but with the headers it can
succeed with 275M of ram (although it takes 5.5mins).