Fix the `context` dependency error after allowing go 1.7.
With govendor no longer blacklisting the go17 build tag, it pulls in
vendor/golang.org/x/net/context/go17.go, which depends on that "context"
package that is supplied by go 1.7. This file is not built under go 1.6
due to the build tag, but govendor is not aware of this, and hence under
1.6 it lists "context" as a missing dependency. This change specifically
excludes just this case.

("golang.org/x/net/context" was moved into the standard library as
"context" in go 1.7, hence this problem).