First steps towards the new SCIOND API (#1002)
This PR defines the new SCIOND API (issue #986) and implements the plumbing for path requests and replies to and from SCIOND.
To that end, I introduced a new entity FwdPathMeta that contains a forwarding path (currently only a SCIONPath object is supported), a list of interfaces of the path and the MTU. The interfaces and MTU have been moved out of SCIONPath. FwdPathMeta is exclusively used by SCIOND and applications, while SCIONPath is the forwarding path contained in SCION packets. This change necessitated the move of PathCombinator into a separate file (which was overdue anyway).