Switch to circleci 2.0
This is a major overhaul of how we build and test the code in a CI environment.
The biggest change is that we now supply a docker image (`scion_base`) that is
run directly by circleci. This image contains all the dependencies of the code,
but not the code itself.

One of the big advantages of circle 2.0 is that there is now a way to run
locally using the ~same environment as circleci.com uses; `tools/ci/local`
automates this, using the "local" build target in `.circleci/config.yml` and
the `scion` docker image.

Notes:
- ci 2.0 stops processing as soon as any step exits with a non-zero status. As
  we have multiple independent tests to run, `tools/ci/run` is used to record
  failures for non-critical steps, and `tools/ci/post_check` is run at the end
  to report any failures.
- `deps.sh` has been split up into `env/*`. This allows go/python/etc
  dependencies to be handled completely seperately from each other. The old
  `./deps.sh all` is now `./env/deps`; the old `./deps.sh golang` is now
  `./env/go/deps`, and so on.
- `./docker.sh base` builds `scion_base`, which is used directly by
  circleci.com. It contains all the dependencies, but no code.
- `./docker.sh build` builds `scion`, which is based on `scion_base`, and
  includes the code and some basic setup steps like topology generation. This
  is the image that `./docker.sh run` and local circleci use.
- `docker.sh` no longer uses `--privileged` to run docker. This no longer works
  on circleci.com, and the need for it has been removed by a symlink created in
  `scion_base`. This does however mean that mininet won't run in docker
  directly. See `topology/mininet/README.md` for the workaround.
- `docker/deps_check` is run as part of circleci and the `scion` docker image
  build; it ensures that no dependency changes have happened since `scion_base`
  was built.