Check that signatures on capnp objects are versioned correctly.
Signatures for capnp objects used in PCBs are defined based on the
version of the capnp definition used. I.e. which fields are covered by
the signature (and in which order) is specific to a given capnp
definition version. This means that when new fields are added, a new
signature definition (which fields, in what order) should be specified.
While this doesn't matter for a homogenous test-bed, as soon as there
are changes made in a production protocol, backwards compatibility is
critical.

The version of a capnp object is defined to be the highest field number,
for simplicity. This change simply calculates the current version
number, and checks to make sure that the correct signature method is
being called. This will have to change once we start implementing actual
backwards compatibility, but will at least in the meantime catch most
cases of fields being added (or removed) without the signature
definitions being updated.