Don't use docker run --privileged.
As of friday, `docker run --privileged ...` fails on circleci for
unclear reasons. While trying to work around this, i discovered that by
specifying a tmpfs for docker to mount, we don't actually _need_
`--privileged` for normal operation.

NB: the one thing that this breaks is mininet support. This is because
mininet depends on creating and managing cgroups, which isn't allowed
without `--privileged`. To run mininet inside docker, use:

  `DOCKER_ARGS=--privileged ./docker.sh run`