Fix a long-standing subtle timing bug in lib.requests (#1044)
If lib.requests gets requests for X faster than the requests expire, and
if the first query it sends doesn't succeed, no new query is sent. This
has been seen to cause up to ~30s between path queries sent by sciond,
for example.

This patch is a simple but brute-force fix - every request now causes a
query to be sent. In the future we will want more intelligent handling
of repeated queries, but this will at least fix this bug for now.