Simplify handling of dependency changes
Currently, if you change any of the dependency info (e.g. add a package
to `env/debian/pkgs.txt`), circleci will fail during the `build_prep`
stage, and none of the tests will be run. This change adds a `env/x/check`
counterpart to `env/x/deps`. `docker/deps_check` runs each `/check` to
see if dependencies have changed. If they have, it's treated as a
soft-fail on circleci, allowing all the tests to run, and reporting
failure during the `post_check` step. This allows a single PR to update
dependencies and the relevant code. The process looks like this:

- Submit a PR as normal, containing both dependency and code changes.
- When the PR has been LGTM'd, and circleci passes the build apart from
  the dependency soft-fail, a new version of `scion_base` will be built
  from that PR and published.
- The PR is then updated with the new `scion_base` image hash in
  `.circleci/config.yml`, and can be submitted.