Requesting TRCs and certs implemented for BS/PS/sciond. (#1000)
When a pcb or path is received by one of those entities, first all used
trcs/certs in that pcb/path are found. Then it is checked if any of those
trcs/certs are missing. Missing trcs/certs are requested from the sender of
this pcb/path and written to trust store.

This change also removes certificate chains from PCBs, as they are no longer needed.