Fix two type bugs in extension handling in Go
This fixes the intermittant error:
`Number of traceroute hops (0) does not match number of expected interfaces (32)`

This error would only happen when there were 16 interfaces on the path,
which caused an overflow when calculating NumHops.

Also fix a logging statement, and switch pktprint to using
SCIONExtPacket so that extension headers are parsed and printed as well.