Fix handling of reversed one-hop-path packets in the BR.
If a one-hop-path packet causes an error, the reversed packet's
extension will be incorrectly handled by the next router, causing it to
corrupt the hop field. This change adds a check to make sure the hop
field hasn't already been generated, and if it has to skip generating a
new one.

Also:
- Log when interfaces are added/activated/deactivated.
- Add pretty-printing represntation of a revocation info proto object to
  reduce the logging noise.