Fix case where core BS/PS aren't informed about revocations.
The bug can be reproduced by doing:

./test/integration/end2end_test.py 2-24 1-16
./supervisor/supervisor.sh mstop *:br1-13*
./test/integration/end2end_test.py --retries 3 2-24 1-16

That last step should fail to get any path by the third retry, as the
two core links to 1-13 are down. However, as the BS/PS don't receive the
revocations, the BS continues to register segments that passed through
1-13, and the PS continues to hand out these segments.

This change fixes the logic, so that a core AS will notifiy its BS/PS
services of revocations if the destination is a local AS. Running the
above steps with this change causes the second e2e test to (correctly)
fail to get any paths to 1-13.

Also:
- As RevocationInfo now has a `short_desc` method, stop putting them on a new
  line in logging statements.
- Log the `rev_info.short_desc()` in e2e, instead of just the interface.
- Fix tools/zkcleanslate to use the correct topo file name.