Go API for SCIOND (#1115)
* Go API for SCIOND

The API exposes structures that mirror the Cap'n Proto messages for communication with SCIOND. The entire SCIOND reply is passed back to the application. For ease of use, some reply fields (such as ISD-AS pairs) have accessors that convert parts of the reply to the appropriate structs.

Integration tests for this API will be added soon.