Fix memory leak in PS and BS. (#1159)
A flaw in how segment verification was handled when receiving path segments through Zookeeper lead to a memory leak.

On startup of a server it can happen that the first path segments come from Zookeeper. If not all certificates are available the server requests it from the local CS with the cache only flag set. However, the CS doesn't have this certificate yet, which lead to a rerequesting loop. This loop also prevented requesting the certs from the appropriate beacon servers when new path segments were arriving.

Fixed the issue by having newer path segments always overwrite older ones in the unverified segments map.