Add SO_TIMESTAMP support.
With SO_TIMESTAMP on a socket, the kernel returns when a packet arrived
when readmsg is called.

As the size of `struct timeval` isn't available anywhere in the go std
lib, this change adds a very short cgo file to calculate the sizes, and
also handle conversion of `struct timeval` to `time.Time`.

Also:
- Clean up metrics on overlay/conn - it no longer imports the border
  metrics module.
- Add missing subsystem to ringbuf metrics.