Fix a memory leak in br/ringbuf.
When a packet was created by the BR (e.g. via `genPkt()`), references to
this packet would be retained in the egress ringbuf's until overwritten,
preventing it from being GC'd. Over time, this would lead to 1024 64KiB
packet references been kept, causing the RSS to balloon up to a maximum
size of ~200MiB.

This change:
- Clears references from a ringbuf on `Read()`
- Clears references from read/write slices passed to a ringbuf.

Also:
- Reduce the packet buffer size to handle jumbo frames (a 7x reduction),
  as there is currently no reason to support packet sizes beyond that.
- Make RtrPkt.Reset() slightly more comprehensive.