Make PathSegmentTTL and RevocationTreeTTL configurable. (#1164)
This PR adds the ability for each AS to configure the TTL of both their path segments and the revocation tree. The revocation tree TTL has to be at least as large as the path segment TTL.

Since each AS now can have a different revocation tree TTL, each revocation needs to include the TTL of the corresponding tree.