Refactor sig egress to be able to use path policies.
This change also splits up egress into a dispatchers (reads from a TUN
device, and writes to policies ringbuf's) and workers (read from a
policiy's ringbuf, encapsulates the packets, and send them to the remote
SIG).

As packet classifiers and path selection policies aren't merged just
yet, currently there is a single default "policy" that the dispatcher
writes to.

Note:
- Currently the path an egress worker gets is never changed, that
  functionality will be in a following PR.
- The egress code should _probably_ be moved to its own module,
  but that will also require moving PathPolicies and SIGInfo out to
  their own modules, too, at a minimum, so i've decided to not do it
  just yet.