Wrap CtrlPld in a signable outer proto.
This PR is the start of a series to overhaul how signing of protos is
handled in SCION.

- The signature is outside the CtrlPld proto, as we otherwise can't
  exclude it from the verification input.
- This approach marshals the CtrlPld to a []byte and signs that
  representation, as we can't depend on language/platform support for
  canonicalisation of capnp protos, especially with unknown fields (for
  forwards/backwards compatibility).

This will allow us to verify ctrl payload messages and senders in
transit. The signature code is in place, but not used/tested.

Note that this approach implies that if later verification of the
CtrlPld is desired, then the original raw representation must be
retained, as it is not possible for a receiver to reconstruct it
identically.

Also:
- Output cerr.Desc in border/gen.go error logs.
- Copy payloads in ScnPkt.Copy()