Rework PathSegment signing.
This is a continuation of the work in #1310, extending that approach to
cover PathSegment signatures in a forwards/backwards
language/platform-agnostic way.

The python part of the change tries to minimize the churn by (mostly)
hiding the difference between PCB and PathSegment. As the IfID field is
no longer in PathSegment's proto, the beacon server still stores PCBs in
zookeeper.

Also:
- Separate out "PCB" and "PathSegment" concepts and protos. The PCB
  proto is used during beaconing and contains fields that are only used
  during that process. PathSegment is used everywhere else (path
  registration/replies).
- Renamed protos following #803.
- Add enum for segment type (up/down/core).
- build go libs that aren't currently used, to make code breakage
  obvious.
- Switch to 64-bit TRC and Cert versions (matching #1309)
- Fix a bug in scion.sh where python tests failing would stop go tests
  from being run.
- Remove SIBRA support from beacon service/generator, as it hasn't
  worked in a long time, and complicates things.