Fix a logging race-condition. (#1312)
The existing liblog uses a bufio.Writer to buffer log output, with a
goroutine calling `.Flush()` on it every 5s. However, bufio.Writer is
_not_ thread-safe, so if flush was ever called while a logging statement
was called, the results are undefined.

This issue was discovered while investigating why our go processes
sometimes stop logging permenantly (e.g. a week without logging
anything), and need to be restarted. It's not clear if this is related
to the race-condition, only time will tell.