Add support for ReadBatch/WriteBatch socket operations. (#1348)
N.B.: this only works with go 1.9+, and on linux.

Go 1.9 added support for reading and writing batches of packets from
sockets, via golang.org/x/net/ipv4. This patch adds support for this to
overlay/conn, and updates the BR to use it. Some performance testing
shows this gives an extra 4-500MBps performance to the BR.

Also:
- Remove gavv/monotime, now that we're on go1.9 which has monotonic time
  handling.
- Fix major bug in sockopt where supplied values were ignored.