Rework go error nesting. (#1369)
The existing error system does not work well for nested errors, and
tends to cause part of the data to be lost (e.g. merging a CError.Ctx
with the current CError, but losing the Desc field). This makes it hard
to work with, and error-prone.

This patch changes the go codebase to use an interface-based approach to
errors, with the ErrorNest interface as the base. This allows generic
handling of nested errors, and FmtError works with this to allow clear
logging of nested errors.