Fix trc req path bug. (#1370)
If a CS needs to send a trc request to a remote AS, it uses `ISD-0` as
the destination AS for the path lookup, which sciond supports. However,
the current code uses the same IA for the data-plane address, which is
illegal.

This patch adds two helpers to `FwdPathMeta` to allow an app to easily
determine the source/destination ASes, and updates the CS to use this
information.