KeyConfig reload on CS-Go
Enable KeyConfig reload on CS-Go. The reload is started by sending SIG_HUP to the certificate server.
Following is reloaded:
- Trust store
- KeyConfig

The keys are loaded from gen/keys. The keys in KeyConfig are only replaced, if the new keys are usable with the trust files.
Root keys are usable if:
- They are authenticated by an active TRC (newest or in grace period)
 Signing keys are usable if:
 - They are authenticated by the newest certificate chain (*)
 - They authenticating certificate chain is verifiable with active TRC (newest or in grace period)
(*) This is necessary since the core AS issuing leaf certificates has a mapping [AS->verifying key] and will only accept reissuance requests if the key matches.
Thus having a different key (even if authenticated by an older but still active certificate chain) means the certificate server is not fully operational.