Add customer mapping
Add customer mapping during topology generation.

Core ASes are responsible for all ASes they issue a certificate for.
Core ASes store verifying keys associated with their customers to verify certificate reissuance requests.
The naming convention is ISDx-ASy-Vz.key, where z denotes the first certificate version which authenticates the key.
For each AS, the key with the highest associated version is loaded.
New key files are only written, when the verifying key changes.

If the core AS needs to remove a customer or change the verifying key manually (e.g. key compromise), the key files
are adjusted as desired and a SIG_HUP is sent to the CS-go process to force reload the customer mapping.

 Additionally:
  - move trust store from cert_srv/main to conf
  - wrap keyConfig in functions to guarantee synchronized acces