Include address and path header lengths in SIG frame size calculation. (#1388)
The current SIG code assumes that the address+path headers are <= 40B
long. This only happens to work because the existing environmets have a
single hop involving a core AS, and using IPv4 addresses. This fixes the
SIG to calculate the actual address+path header sizes, and use those.

N.b. it still does not include any extensions that might be in use. That
will eventually depend on snet providing that information.