Change addr.ISD_AS to value and rename to addr.IA
Currently most of the code uses *addr.ISD_AS despite not needing to.
This leads to bugs which are very painful to debug where an IA is
inadvertently changed or reused after being shared.
There are very few places in the code where the fact that *addr.IA is a
pointer rather than a value is relied upon (the BR is the main
exception), and these are fairly easy to work around.

The rename is due to popular demand :)