Change SIG to only use a single TUN device.
This PR adds a new goroutine, egress.Reader. It is responsible for:
- reading packets from the TUN device
- consulting egress.NetMap to determine the correct ringbuf to use
- writing packets to the ringbuf

egress.Networks is a basic implementation of the egress.NetMapI
interface, which defines how an IP can be mapped to an allocation
belonging to a remote AS.

As the sig now only needs CAP_NET_ADMIN at start, capabilities are
cleared once the TUN device is opened and routes injected.

Also:
- Rename sig.go to main.go for consistency.