IFID messages hop-by-hop
This PR fixes a problem introduced by #1476.
IFID msgs are now sent from BS to local BR and from there forwarded to the remote BR which forwards it to the local BS.
Like that IFID msgs will not generate SCMP revocations at the BR if the interface has recently come up.