path_server: handle pending request after PCB processing is complete
Currently there is a race condition between Path Server main thread and
worker thread. When a sciond queries a local PS for a path from a non
core AS to a core AS. Local PS forwards that request to core PS for the
core segments and starts populating PathSegmentDB. While the main thread
is busy processing that PCB, worker thread handles pending request and
replies with whatever present inside PathSegmentDB. Usually it is just
one path and rest of the possible paths are neglected.

For example:

```
./bin/showpaths -dstIA 2-ff00:0:210 -srcIA 1-ff00:0:133
INFO[05-27|09:38:27] Started                                  id=df259bab goroutine=dispatcher_bck
Available paths to 2-ff00:0:210
[ 0] Hops: [1-ff00:0:133 86>96 1-ff00:0:132 39>73 1-ff00:0:131 91>95 1-ff00:0:130 59>76 1-ff00:0:120 77>55 1-ff00:0:110 57>19 2-ff00:0:210] Mtu: 1280
```

There are 4 possible paths from 133 to 210 but it only returns one. An
easy way to verify that there are more path just flush the cache of
sciond and query again for the same path and you will find that there
are more paths.