Fix replies from BR to BR in the same AS
The OverlayPort in the metadata information when reading packets from
an internal interface was always set to the endhost overlay port.

When a BR uses this information to generate a reply it would use the
wrong overlay port.

Fix this problem by setting the source overlay port to the L4 port of
the received packet.