Introduce new interfaces for Address types
This change introduces two new interfaces AppAddr and OverlayAddr.
Both have a similar interface but the work with different Address Types,
ie. SVC as an AppAddr.

Internally all current objects implementing the interfaces consist of
a HostAddr and a port.

Only allowing to create objects that implement these interfaces through
constructors, we can restrict which HostAddr as valid for each
interface.