Comments r1
Introduce new L4 address interface

A new L4 address interface (L4Info), along with the L3/HostAddr interface,
is used as the base for more complex addresses, such as AppAddr or
OverlayAddr.

AppAddr allows for any combination of HostAddr + L4Info, while
OverlayAddr is currently restricted to IP + UDP.

To achieve this restriction, the fields in OverlayAddr struct are
private and a constructor is used to create structs of such type.