snet: do not ignore nexthop address when provided
Currently the nexthop address when writting to a destination in the
local AS is ignored.

With this change the nexthop address will be used regardless of the
destination being in the local AS or not, while still having the same
functionality in the case of not specifying a nexthop.
Also returns an error if a path without a nexthop has been provided.

Fixes #1503