Don't use sudo -v
scionproto/scion#1787 used `sudo -v` as a way to ensure that sudo
credentials are cached before running the desired commands with sudo.
This, however, breaks in an environment where the user in question can
run any command without a password, as `sudo -v` will always prompt for
a password if no sudo credentials are present.

To work around this, this PR changes from using `sudo -vp <prompt>` to
`sudo -p <prompt> true`, which will actually run a (harmless) command.