sciond: Increase timeout for PathReply processing
On a CI env we have seen that processing 7 segments can take more than 200ms.
Since the last PCB was the missing piece for the path,
sciond wrongly answered with 0 path, even though
there would have been one if it waited longer.

Increasing the timeout is a temporary fix for this.
We expect the go-sciond to be better in this regard.