Fix writing of offsets in scion packets (go) (#1810)
It could be that the s.Path is not nil but empty, then we should also not write CurreInfoF and CurrHopF.
These fields will have the same value (0, i.e. offset / common.LineLen) which will lead to a parse error.
Also guard GetInfoF and GetHopField methods against an empty path.